avformat/wtvdec: Use 64bit for ret to avoid overflow
(cherry picked from commit d44e0d8b930732a4a247b4884d75cf62b4ad3664)

+++ libavformat/wtvdec.c
@@ -760,7 +760,7 @@ static int recover(WtvContext *wtv, uint64_t broken_pos)
     int i;
     for (i = 0; i < wtv->nb_index_entries; i++) {
         if (wtv->index_entries[i].pos > broken_pos) {
-            int ret = avio_seek(pb, wtv->index_entries[i].pos, SEEK_SET);
+            int64_t ret = avio_seek(pb, wtv->index_entries[i].pos, SEEK_SET);
             if (ret < 0)
                 return ret;
             wtv->pts = wtv->index_entries[i].timestamp;
